refactor: move  queries to Commands folder with validators
