+++ RO.DevTest.Application/Features/Product/Commands/GetPagedProductsCommand/GetPagedProductResult.cs
@@ -1,6 +1,6 @@
-﻿namespace RO.DevTest.Application.Features.Product.Queries.GetPagedProducts;
+﻿namespace RO.DevTest.Application.Features.Product.Commands.GetPagedProductsCommand;
 
-public class ProductResult
+public class GetPagedProductResult
 {
     public string Id { get; set; } = string.Empty;
     public string Name { get; set; } = string.Empty;
