+++ RO.DevTest.Application/Features/Product/Commands/GetPagedProductsCommand/GetPagedProductsCommandValidator.cs
@@ -1,6 +1,6 @@
 ﻿using FluentValidation;
 
-namespace RO.DevTest.Application.Features.Product.Queries.GetPagedProducts;
+namespace RO.DevTest.Application.Features.Product.Commands.GetPagedProductsCommand;
 
 public class GetPagedProductsCommandValidator : AbstractValidator<GetPagedProductsCommand>
 {
