+++ RO.DevTest.Application/Features/Sale/Commands/GetPagedSales/GetPagedSalesCommandValidator.cs
@@ -1,6 +1,6 @@
 ﻿using FluentValidation;
 
-namespace RO.DevTest.Application.Features.Sale.Queries.GetPagedSales;
+namespace RO.DevTest.Application.Features.Sale.Commands.GetPagedSales;
 
 public class GetPagedSalesCommandValidator : AbstractValidator<GetPagedSalesCommand>
 {
