+++ RO.DevTest.Application/Features/Sale/Commands/GetPagedSales/SaleItemResult.cs
@@ -1,4 +1,4 @@
-﻿namespace RO.DevTest.Application.Features.Sale.Queries.GetPagedSales;
+﻿namespace RO.DevTest.Application.Features.Sale.Commands.GetPagedSales;
 
 public class SaleItemResult
 {
